requirements.txt
# Core
pandas
numpy
pydantic
pyyaml
sqlalchemy
websocket-client

# Dev/Test
pytest
